// src/api/auth/model/modelsUserRole.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * Auth API
 * Authentication and authorization service for the PRPO project
 * OpenAPI spec version: 1.0
 */

export type ModelsUserRole = (typeof ModelsUserRole)[keyof typeof ModelsUserRole]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ModelsUserRole = {
  RoleCustomer: 'customer',
  RoleEmployee: 'employee',
  RoleAdmin: 'admin',
} as const