// src/api/auth/model/apiUserResponse.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * Auth API
 * Authentication and authorization service for the PRPO project
 * OpenAPI spec version: 1.0
 */
import type { ModelsUserRole } from './modelsUserRole'

export interface ApiUserResponse {
  active?: boolean
  created_at?: string
  email?: string
  first_name?: string
  id?: string
  last_name?: string
  role?: ModelsUserRole
  updated_at?: string
}